+++ packages/next-common/components/charts/globalConfig.js
@@ -1,0 +1,12 @@
+import {
+  Chart,
+  LineElement,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  Tooltip,
+} from "chart.js";
+
+Chart.defaults.font.family = "Inter";
+
+Chart.register(LineElement, CategoryScale, LinearScale, PointElement, Tooltip);